pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "discord-webp-tuner"
version = "0.1.0"
description = "WebP parameter sweep + quality metrics tuned for Discord downscaled viewing."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "Pillow>=10.0",
  "numpy>=1.26",
  "pandas>=2.0",
  "matplotlib>=3.8",
  "tqdm>=4.66",
  "typer>=0.12",
]

[project.optional-dependencies]
metrics = [
  "torch",
  "piq>=0.8",
  "ssimulacra2>=0.3.0",
]
interactive = [
  "plotly>=5.18",
]
dev = [
  "pytest>=7.0",
]

[project.scripts]
discord-webp-tuner = "discord_webp_tuner.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]
